Refactor: Merge compiler and AST modules
