+++ internal/quipple/compiler/tree_nodes.go
@@ -1,4 +1,4 @@
-package ast
+package compiler
 
 import (
 	"qp/internal/pkgdata"
